--- include/applications/poffin_case/manager_main.h
@@ -1,10 +1,0 @@
-#ifndef POKEPLATINUM_APPLICATIONS_POFFIN_CASE_MANAGER_MAIN_H
-#define POKEPLATINUM_APPLICATIONS_POFFIN_CASE_MANAGER_MAIN_H
-
-#include "overlay_manager.h"
-
-BOOL PoffinManager_Init(ApplicationManager *appMan, int *unused);
-BOOL PoffinManager_Main(ApplicationManager *appMan, int *unused);
-BOOL PoffinManager_Exit(ApplicationManager *appMan, int *unused);
-
-#endif // POKEPLATINUM_APPLICATIONS_POFFIN_CASE_MANAGER_MAIN_H
