+++ src/meson.build
@@ -326,7 +326,7 @@ pokeplatinum_c = files(
     'pokedex_language.c',
     'pokedex_heightweight.c',
     'pokedex_data_index.c',
-    'poffin_case.c',
+    'applications/poffin_case/main.c',
     'poffin_sprite.c',
     'unk_02099500.c',
     'unk_02099550.c',
@@ -815,7 +815,7 @@ pokeplatinum_c = files(
     'game_opening/ov77_021D6670.c',
     'game_opening/ov77_021D6C70.c',
     'choose_starter/choose_starter_app.c',
-    'applications/poffin_case/manager_main.c',
+    'applications/poffin_case/manager.c',
     'applications/poffin_case/menus.c',
     'applications/poffin_case/sprites.c',
     'applications/poffin_case/ov79_021D2268.c',
