+++ src/unk_0203D1B8.c
@@ -14,7 +14,6 @@
 #include "struct_decls/struct_0209747C_decl.h"
 #include "struct_defs/choose_starter_data.h"
 #include "struct_defs/gts_player_data.h"
-#include "struct_defs/poffin_case.h"
 #include "struct_defs/struct_0203DA00.h"
 #include "struct_defs/struct_0203DDFC.h"
 #include "struct_defs/struct_0203DE34.h"
@@ -39,6 +38,7 @@
 #include "applications/pc_boxes/box_app_manager.h"
 #include "applications/pc_boxes/pokemon_storage_session.h"
 #include "applications/pc_hall_of_fame/manager.h"
+#include "applications/poffin_case/main.h"
 #include "applications/pokedex/pokedex_main.h"
 #include "applications/pokemon_summary_screen/main.h"
 #include "applications/town_map/main.h"
@@ -99,7 +99,6 @@
 #include "overworld_map_history.h"
 #include "party.h"
 #include "player_avatar.h"
-#include "poffin_case.h"
 #include "pokedex.h"
 #include "pokemon.h"
 #include "poketch.h"
